+++ NEWS.md
@@ -15,4 +15,7 @@ FEATURES / CHANGES
   transformations
 - add ability to set `options(defaultLocalDiskControl = ...)`, etc. so that you 
   do not always need to specify `control=` in all MapReduce-inducing operations
+- add print method for key-value pairs to show things nicely, particularly
+  when the object is a ddf, only show top rows of value
+- add labels "key" and "value" to key-value pairs